Renamed routes and controllers to MOTD
