+++ apps/express-backend/src/routes/motd.routes.ts
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import * as controller from "../controllers/index.controllers.js";
+import * as controller from "../controllers/motd.controllers.js";
 
 const indexRouter: Router = Router();
